requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
auth0-python==4.7.2
babel==2.16.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
# Editable install with no version control (bigbets_origination_crews==0.1.0)
-e /teamspace/studios/this_studio/bigbets_origination_crews
bleach==6.2.0
bokeh==3.6.1
boto3==1.35.54
botocore==1.35.54
Brotli==1.1.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.18
click==8.1.7
cohere==5.11.4
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
crewai==0.86.0
crewai-tools==0.17.0
cryptography==43.0.3
cssselect2==0.7.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
# Editable install with no version control (demand_signals_crew==0.1.0)
-e /teamspace/studios/this_studio/demand_signals_crew
Deprecated==1.2.15
deprecation==2.1.0
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
docx2txt==0.8
durationpy==0.9
embedchain==0.1.125
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.4
fastavro==1.9.7
fastjsonschema==2.20.0
filelock==3.16.1
fire==0.7.0
flatbuffers==24.3.25
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.10.0
google-api-core==2.23.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.72.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
gptcache==0.1.44
graphviz==0.20.3
greenlet==3.1.1
grpc-google-iam-v1==0.13.1
grpcio==1.68.0
grpcio-status==1.68.0
grpcio-tools==1.68.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
instructor==1.6.4
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
json_repair==0.30.2
jsonpatch==1.33
jsonpickle==4.0.0
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
kubernetes==31.0.0
lancedb==0.16.0
langchain==0.3.7
langchain-cohere==0.3.1
langchain-community==0.3.7
langchain-core==0.3.19
langchain-experimental==0.3.3
langchain-openai==0.2.8
langchain-text-splitters==0.3.2
langsmith==0.1.143
lightning==2.4.0
lightning-cloud==0.5.70
lightning-utilities==0.11.8
lightning_sdk==0.1.30
linkify-it-py==2.0.3
litdata==0.2.19
litellm==1.52.9
litserve==0.2.3
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
markdown_tables_extended==0.1.3
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.8.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mem0ai==0.1.29
mistune==3.0.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nodeenv==1.9.1
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.20.0
openai==1.54.4
openpyxl==3.1.5
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
orjson==3.10.11
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.1.4
pandocfilters==1.5.1
panel==1.5.4
param==2.1.1
parameterized==0.9.0
parso==0.8.4
pdfkit==1.0.0
pdfminer.six==20231228
pdfplumber==0.11.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
portalocker==2.10.1
posthog==3.7.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydyf==0.11.0
Pygments==2.18.0
PyJWT==2.9.0
pylance==0.19.2
pyparsing==3.2.0
pypdf==5.1.0
pypdfium2==4.30.0
pyphen==0.15.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyright==1.1.389
pysbd==0.3.4
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.17
pytorch-lightning==2.4.0
pytube==15.0.0
pytz==2024.2
pyvis==0.3.2
pyviz_comms==3.0.3
PyYAML==6.0.2
pyzmq==26.2.0
qdrant-client==1.12.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
# Editable install with no version control (research_crew==0.1.0)
-e /teamspace/studios/this_studio/research_crew
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.20.1
rsa==4.9
s3transfer==0.10.3
schema==0.7.7
scikit-learn==1.3.2
scipy==1.11.4
selenium==4.26.1
Send2Trash==1.8.3
shapely==2.0.6
shellingham==1.5.4
simple-term-menu==1.6.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.35
stack-data==0.6.3
starlette==0.41.2
# Editable install with no version control (supply_signals_crew==0.1.0)
-e /teamspace/studios/this_studio/supply_signals_crew
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
termcolor==2.5.0
terminado==0.18.1
# Editable install with no version control (test_multi_crew==0.1.0)
-e /teamspace/studios/this_studio/test_multi_crew
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.20.3
tomli==2.0.2
tomli_w==1.1.0
torch==2.2.1+cu121
torchmetrics==1.3.1
torchvision==0.17.1+cu121
tornado==6.4.1
tqdm==4.66.6
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
triton==2.2.0
typer==0.13.0
types-python-dateutil==2.9.0.20241003
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
uri-template==1.3.0
urllib3==2.2.3
uv==0.5.2
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
wcwidth==0.2.13
weasyprint==63.0
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.1.2
# Editable install with no version control (whitespace_identification_crew==0.1.0)
-e /teamspace/studios/this_studio/whitespace_identification_crew
widgetsnbextension==4.0.13
wrapt==1.16.0
wsproto==1.2.0
xyzservices==2024.9.0
yarl==1.17.1
zipp==3.21.0
zopfli==0.2.3.post1
